HFX-2158 CA-285840: Make sure stunnel doesn't hang talking to dead hosts
It would be way better to be using select rather than blocking
reads, but that's a much bigger change and this will need to be
backported.

Also there's a new mechanism for breaking out of the retry loop
if a host has been marked as dead, which will require a change
in xapi to be used.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>
Signed-off-by: Christian Lindig <christian.lindig@citrix.com>