CA-340493: only call diagnose failure if a logfile exists
A logfile is only retained if extended_diagnosis is enabled when
connecting Stunnel.
The xe CLI only enables extended diagnosis when debugonfail is set,
which is not the default. Thus a connection reset in the xe CLI would
incorrectly fail with a Sys_error about not being able to find the
stunnel logfile.
Now it would reraise the Connection_reset and allow that to be printed
by the toplevel exception handler.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>