REQ-403 CP-33822 make cert checking a global flag
We either enable cert checking everywhere, or nowhere at all. The idea
is that xapi will enable it using `Stunnel.set_verify_tls_cert`.

This simplifies calls to `with_connect`, since we no longer need a
`verify_cert` parameter.

We still need a way to remove all stunnels, regardless of whether they
are verified or not. This has been implemented with a `try_all`
parameter.

Signed-off-by: lippirk <ben.anson@citrix.com>