CP-36100 for SNI xapi:pool use pool cert bundle
This concerns the configurations of clients that use Stunndel to connect
to a host. If the connection demands the connection to be verified, a
configuration using SNI 'pool' is used. This requires the client to know
the certificates of all hosts in the pool. The configuration expects to
find them in a single file that bundles them:

  /etc/stunnel/xapi-pool-ca-bundle.pem

The construction of this bundle is not the responsibility of this
library, however.

All this patch does is to make sure the correct bundle is used in the
case a verified connection is opened.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>