xapi-compression: change misleading error message
Previously it always stated zcat file while ignoring the actual name of
the executable used. Now it's taken into account when generating the
error. The error is now logged instead of output to stderr.

This also contains tweaks to the dune metadata, making it less reliable
on implicit dependencies. The merlin approximate mode is not needed
because .merlin files aren't generated anymore.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>