CA-368579: Limit concurrent connections with semaphore
To add some protection against overloading the server, and running out
of file descriptors and other resources, the number of concurrent
connections is now limited to 800 by default. Connections coming in when
the limit has been reached are put on hold, and not accepted until
another connection finishes.

There are three separate buckets for connections to the main unix
socket, the client-cert (unix) socket and the TCP socket.

This also add some more debug logging when connections are accepted and
disconnected. This is disabled by default along with other "http" logs,
but can be enabled in the conf file.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>

(cherry picked from commit 38998bf272b9fe9ec5996b460a43225cc0e744d7)